Create CreateCartItemRequest class and rename previous request class to CreateCartItemDto
